+++ contracts/interfaces/ILeverageAdapter.sol
@@ -13,7 +13,7 @@ interface ILeverageAdapter is IController {
         uint256 collateralAmount,
         uint256 borrowAmount,
         uint256 maxPayInterest,
-        uint8 loanEpochs,
+        uint16 loanEpochs,
         bytes memory swapData,
         ERC20PermitParams calldata collateralPermitParams
     ) external payable;
